# openapi/openapi-contract-consumption-be-tier1-v1.yaml
info:
  title: Contract Consumption API
  description: |-
    This API enables the discovery, negotiation, and consumption of data contracts
    between data providers and consumers in a secure and standardized way. It supports transfer initiation,
    contract negotiation, and resource address management.

    ## Business Context

    The API facilitates data exchange in a federated data space. Consumers can search for data offers, negotiate contracts,
    and initiate data transfers from providers. The API ensures secure and traceable interactions between participants.


    ### Key Actors

    - **Provider**: Offers data assets.
    - **Consumer**: Requests and consumes data.

    ### Typical Flow

    1. Search for data offers.

    2. Negotiate a contract.

    3. Initiate a transfer.

    4. Monitor transfer status.
  contact:
    name: Simpl Programme
    url: https://simpl-programme.ec.europa.eu/
    email: cnect-simpl@ec.europa.eu
  license:
    name: European Union Public License (EUPL) 1.2
    url: https://interoperable-europe.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_en.pdf
  version: "1.0"
servers:
- url: https://contract-consumption-be.dev.simpl-europe.eu/v1
security:
- bearerAuth: []
openapi: 3.0.1
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: IAA cloud gateway JWT token
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Template:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
    InputObject:
      type: object
      properties:
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
...
    InputValidationIssue:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/InputObject'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    InternalServerErrorProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    NotFoundProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    BadRequestProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        issues:
          type: array
          items:
            $ref: '#/components/schemas/InputValidationIssue'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
paths:
...
  /resourceAddresses/templates/{templateId}/schema:
    get:
      tags:
      - Resource Address
      summary: Returns a destination address template schema based on templateId
      description: Returns the JSON schema template for a specific resource address. This schema defines the structure and constraints for configuring a data destination.
      operationId: getDestinationAddressTemplate
      parameters:
      - name: templateId
        in: path
        description: The template id
        required: true
        schema:
          type: string
          example: TPL-1
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: Template (JSON schema) for the sourceAddress
          content:
            application/json:
              schema:
                type: object
              examples:
                IonosS3:
                  summary: IonosS3 template example
                  description: IonosS3
                  value:
                    type: object
                    properties:
                      bucketName:
                        type: string
                      region:
                        type: string
                    required:
                    - bucketName
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                Not Found Error Example:
                  description: Not Found Error Example
                  value:
                    type: urn:problem-type:simpl:resourceAddressNotFound
                    title: Resource Address Not Found
                    status: 404
                    detail: No template found for templateId 1
...
  /resourceAddresses/sharingMethods/{sharingMethodId}/templates:
    get:
      tags:
      - Resource Address
      summary: Returns the list of destination address templates by sharing method and offering type.
      description: Fetches a list of address templates associated with a specific sharing method and offering type. Helps consumers understand how to configure data destinations.
      operationId: getDestinationAddressTemplates
      parameters:
      - name: offeringType
        in: query
        description: The offering type identifier
        required: true
        schema:
          type: string
          enum:
          - APPLICATION
          - DATA
          - INFRASTRUCTURE
      - name: sharingMethodId
        in: path
        description: The resource sharing method identifier
        required: true
        schema:
          type: string
          enum:
          - HTTPDATA_PUSH
          - IONOS_S3
          - REST_API
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: A list of templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
              examples:
                templates:
                  summary: List of templates
                  description: templates
                  value:
                  - id: "1"
                    label: Template 1
                  - id: "2"
                    label: Template 2
                  - id: "3"
                    label: Template 3
  /resourceAddresses/templates/{templateId}/uiSchema:
    get:
      tags:
      - Resource Address
      summary: Returns a destination address UI schema based on templateId.
      description: Returns a UI schema for rendering the resource address form. This helps front-end applications generate user-friendly forms for data destination configuration.
      operationId: getDestinationAddressUiSchema
      parameters:
      - name: templateId
        in: path
        description: The template id
        required: true
        schema:
          type: string
          example: TPL-1
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "200":
          description: UI schema for the sourceAddress
          content:
            application/json:
              schema:
                type: object
              examples:
                amazonS3UI:
                  summary: UI schema example
                  description: amazonS3UI
                  value:
                    ui:order:
                    - bucketName
                    - region
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                Not Found Error Example:
                  description: Not Found Error Example
                  value:
                    type: urn:problem-type:simpl:resourceAddressNotFound
                    title: Resource Address Not Found
                    status: 404
                    detail: No UI schema found for templateId 1
...